Refactor utils.
Separate musicPlayer utils.
